refactor(cli): extract JSM-create into src/cli/issue/jsm_create.rs (Seam A)
Behavior-preserving shard per the 2026-06-25 architecture analysis (refactor-2026-06-25/, DO-PARTIAL verdict, Seam A). Moves handle_jsm_create + resolve_jsm_request_type_id + JsmCreateArgs verbatim out of create.rs (2,880→2,447 LOC) into a new jsm_create.rs (444 LOC). No logic change. Invariant I-1 (ADR-0014 request_type.is_some() dispatch fork) byte-for-byte intact. Test counts unchanged (1957 passed / 93 ignored before and after — no test added or dropped). Updated CLAUDE.md src-tree + create.rs size-deviation; fixed 2 stale rustdoc citations in tests/issue_create_jsm.rs. ADR-0012 shard.
